fix: avoid hang when exporting voice chat (#262)
* work on #261: avoid hang

See #261 for the situation. This commit avoids errors caused by multimodal parts we don't understand.

It does NOT add any functionality that understands advanced voice mode parts. Perhaps that should be broken out of #261 into a separate enhancement request.

* address es-lint complaint

eslint error:

error '&&' should be placed at the beginning of the line

solution:

move a trailing `&&` to the front of the next line